requirements.txt
python-dotenv
loguru
pipecat-ai[openai,silero,daily]
